merapikan baris kode

+++ lib/pages/formPage.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: file_names
 
-import 'package:baca_buku_flutter/pages/home1.dart';
+import 'package:baca_buku_flutter/pages/home.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:flutter/material.dart';
 
@@ -15,10 +15,7 @@ class FormPage extends StatefulWidget {
 }
 
 class _FormPageState extends State<FormPage> {
-  //set form key
   final _formKey = GlobalKey<FormState>();
-
-  //set texteditingcontroller variable
   var gambarController = TextEditingController();
   var judulController = TextEditingController();
   var penulisController = TextEditingController();
@@ -26,25 +23,14 @@ class _FormPageState extends State<FormPage> {
   var sinopsisController = TextEditingController();
   var halamanController = TextEditingController();
 
-  //inisialize firebase instance
   FirebaseFirestore firebase = FirebaseFirestore.instance;
   CollectionReference? books;
 
   void getData() async {
-    //get users collection from firebase
-    //collection is table in mysql
     books = firebase.collection('books');
-
-    //if have id
     if (widget.id != null) {
-      //get users data based on id document
       var data = await books!.doc(widget.id).get();
-
-      //we get data.data()
-      //so that it can be accessed, we make as a map
       var item = data.data() as Map<String, dynamic>;
-
-      //set state to fill data controller from data firebase
       setState(() {
         gambarController = TextEditingController(text: item['gambar']);
         judulController = TextEditingController(text: item['judul']);
@@ -58,8 +44,6 @@ class _FormPageState extends State<FormPage> {
 
   @override
   void initState() {
-    // ignore: todo
-    // TODO: implement initState
     getData();
     super.initState();
   }
@@ -79,17 +63,37 @@ class _FormPageState extends State<FormPage> {
           backgroundColor: Colors.white,
           foregroundColor: Colors.deepPurple,
           actions: [
-            //if have data show delete button
             widget.id != null
                 ? IconButton(
                     onPressed: () {
-                      //method to delete data based on id
-                      books!.doc(widget.id).delete();
-
-                      //back to main page
-                      // '/' is home
-                      Navigator.of(context).pushNamedAndRemoveUntil(
-                          '/one', (Route<dynamic> route) => false);
+                      showDialog(
+                          context: context,
+                          builder: (BuildContext ctx) {
+                            return AlertDialog(
+                              title: const Text('Konfirmasi'),
+                              content: const Text(
+                                  'Apakah kamu yakin ingin menghapus ini?'),
+                              actions: [
+                                TextButton(
+                                    onPressed: () {
+                                      books!.doc(widget.id).delete();
+                                      Navigator.push(
+                                        context,
+                                        MaterialPageRoute(
+                                            builder: (context) => HomePage(
+                                                  email: widget.email,
+                                                )),
+                                      );
+                                    },
+                                    child: const Text('Iya')),
+                                TextButton(
+                                    onPressed: () {
+                                      Navigator.of(context).pop();
+                                    },
+                                    child: const Text('Tidak'))
+                              ],
+                            );
+                          });
                     },
                     icon: const Icon(
                       Icons.delete,
@@ -98,8 +102,6 @@ class _FormPageState extends State<FormPage> {
                 : const SizedBox()
           ],
         ),
-        //this form for add and edit data
-        //if have id passed from main, field will show data
         body: Form(
           key: _formKey,
           child: ListView(padding: const EdgeInsets.all(16.0), children: [
@@ -235,8 +237,6 @@ class _FormPageState extends State<FormPage> {
               ),
               onPressed: () {
                 if (_formKey.currentState!.validate()) {
-                  //if id not null run add data to store data into firebase
-                  //else update data based on id
                   if (widget.id == null) {
                     books!.add({
                       'gambar': gambarController.text,
@@ -258,16 +258,12 @@ class _FormPageState extends State<FormPage> {
                       'email': widget.email
                     });
                   }
-                  //snackbar notification
                   const snackBar = SnackBar(
                     content: Text('Buku berhasil diunggah!',
                         style: TextStyle(color: Colors.black)),
                     backgroundColor: Colors.amber,
                   );
                   ScaffoldMessenger.of(context).showSnackBar(snackBar);
-
-                  //back to main page
-                  //home page => '/'
                   Navigator.pushReplacement(
                       context,
                       MaterialPageRoute(
